restructure ... prep for introducing router ...
