+++ app/assets/app/components/header-nav/index.ts
@@ -1,0 +1,1 @@
+export * from './header-nav.component'
