+++ app/assets/app/components/search-management/index.ts
@@ -1,0 +1,1 @@
+export * from './search-management.component'
